server/requirements.txt
flask 
pymongo
flask-cors
flask-bcrypt 
mysql-connector-python
Flask-Session
python-dotenv
redis